backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
Pillow>=10.3.0

# NumPy (must be < 2.0 for PyTorch compatibility)
numpy<2.0

# Machine Learning - Stable Diffusion
torch>=2.6.0
torchvision>=0.19.0
diffusers>=0.27.0
transformers>=4.57.1
accelerate>=0.27.0
huggingface-hub>=0.20.0

# ControlNet Preprocessing
controlnet-aux==0.0.7

# Utilities
python-dotenv==1.0.0

